LPS-74839 petra-json-web-service-client - Reorganize - extract all reusable code to base abstract class
